refactor(realtime): replace SSE with WebSocket + Long Polling
BREAKING CHANGE: SSE endpoints removed, WebSocket is now primary transport

## Changes

### Backend
- Add budget_ws.py: WebSocket manager, endpoints, broadcast functions
- Add Long Polling fallback endpoint (/api/v1/budget/poll)
- Remove budget_sse.py and all SSE-related code
- Update facts.py, transfers.py, shopping_list_items.py:
  - Import from budget_ws module
  - Rename sse -> ws variable naming for consistency
- Update router.py and main.py for WS lifecycle

### Frontend
- Add budgetWSClient.js (1423 lines) with:
  - WebSocket as primary transport (bidirectional)
  - Long Polling fallback (10 sec interval)
  - Multi-tab support (Web Locks + BroadcastChannel)
  - Safari iOS per-tab mode (no Web Locks)
  - Backward compatibility (window.budgetSSEClient alias)
  - Status indicator badge (8 states)
- Remove budgetSSEClient.js
- Update base.html to use new client

### Nginx
- Add WebSocket location block with proper headers
- Remove SSE location block

### Documentation
- Create websocket.yaml and ws-broadcast.yaml
- Update all architecture docs (30+ files) SSE -> WebSocket
- Remove sse.yaml and sse-broadcast.yaml

## API Changes
- WebSocket: wss://domain/api/v1/budget/ws?token=<JWT>
- Long Polling: GET /api/v1/budget/poll?since=<ts>&timeout=10
- Bidirectional: ping/pong, check_online messages
- Same event types: fact_created, fact_updated, etc.
