+++ backend/app/api/v1/endpoints/transfers.py
@@ -18,17 +18,17 @@ logger = logging.getLogger(__name__)
 router = APIRouter(prefix="/transfers", tags=["Transfers"])
 
 
-# Lazy import for SSE broadcast to avoid circular dependencies
-_budget_sse_module = None
+# Lazy import for WebSocket broadcast to avoid circular dependencies
+_budget_ws_module = None
 
 
-def _get_budget_sse_broadcast():
-    """Get budget SSE module lazily to avoid circular imports."""
-    global _budget_sse_module
-    if _budget_sse_module is None:
-        from backend.app.api.v1.endpoints import budget_sse
-        _budget_sse_module = budget_sse
-    return _budget_sse_module
+def _get_budget_ws_broadcast():
+    """Get budget WebSocket module lazily to avoid circular imports."""
+    global _budget_ws_module
+    if _budget_ws_module is None:
+        from backend.app.api.v1.endpoints import budget_ws
+        _budget_ws_module = budget_ws
+    return _budget_ws_module
 
 
 async def generate_transfer_id(session: AsyncSession) -> int:
@@ -231,7 +231,7 @@ async def create_transfer(
 
     # 8. SSE broadcast for transfer created (notify all connected clients)
     try:
-        sse = _get_budget_sse_broadcast()
+        ws = _get_budget_ws_broadcast()
         transfer_data = {
             "id": transfer_id,
             "source_fact_id": expense_fact.id,
@@ -240,7 +240,7 @@ async def create_transfer(
             "transfer_date": str(transfer.transfer_date),
             "description": transfer.description,
         }
-        await sse.broadcast_transfer_created(transfer_data)
+        await ws.broadcast_transfer_created(transfer_data)
     except Exception as e:
         logger.warning(f"SSE broadcast failed for transfer {transfer_id}: {e}")
         # Don't fail the request if broadcast fails
@@ -327,8 +327,8 @@ async def delete_transfer(
 
     # SSE broadcast for transfer deleted (notify all connected clients)
     try:
-        sse = _get_budget_sse_broadcast()
-        await sse.broadcast_transfer_deleted(transfer_id)
+        ws = _get_budget_ws_broadcast()
+        await ws.broadcast_transfer_deleted(transfer_id)
     except Exception as e:
         logger.warning(f"SSE broadcast failed for deleted transfer {transfer_id}: {e}")
         # Don't fail the request if broadcast fails
