+++ frontend/web/templates/base.html
@@ -196,7 +196,7 @@
     {% if user %}
     <!-- Auth-required scripts (только для авторизованных) -->
     <script src="/static/js/offline/pushManager.min.js?v=PLACEHOLDER"></script>
-    <script src="/static/js/budget/budgetSSEClient.js?v=PLACEHOLDER"></script>
+    <script src="/static/js/budget/budgetWSClient.js?v=PLACEHOLDER"></script>
     <script src="/static/js/budget/incrementalUpdates.js?v=PLACEHOLDER"></script>
     {% endif %}
